refactor: metabox factory in order to add more security
