+++ trunk/src/Metabox/Type/class-user.php
@@ -1,8 +1,8 @@
 <?php declare( strict_types = 1 );
 
-namespace Trasweb\Plugins\DisplayMetadata\Metabox;
+namespace Trasweb\Plugins\DisplayMetadata\Metabox\Type;
 
-use Trasweb\Plugins\DisplayMetadata\Plugin;
+use Trasweb\Plugins\DisplayMetadata\Metabox\Metabox;
 
 /**
  * This class manages `Display Metadata` user metabox.
@@ -31,6 +31,10 @@ final class User extends Metabox {
      */
     public function can_be_registered(string $screen_slug): bool
     {
+        if(!parent::can_be_registered($screen_slug)){
+            return false;
+        }
+
         return 'user-edit' === $screen_slug || 'profile' === $screen_slug;
     }
